convert coordinate from 3d camera to ui
