+++ cocos/core/utils/coordinates-converts-utils.ts
@@ -2,18 +2,17 @@
  * @category pipeline
  */
 
-import { UITransformComponent } from '../components/ui-base/ui-transform-component';
-import { CameraComponent } from '../3d/framework';
+import { CameraComponent } from '../3d/framework/camera-component';
 import { Vec3 } from '../math';
 import { Node } from '../scene-graph';
-
-const _temp_vec3_1 = new Vec3();
+import { replaceProperty } from './deprecated';
 
 /**
  * @en
  * Conversion of non-UI nodes to UI Node (Local) Space coordinate system.
  * @zh
  * 非 UI 节点转换到 UI 节点(局部) 空间坐标系。
+ * @deprecated 将在 1.2 移除，请使用 CameraComponent 的 `convertToUINode`。
  * @param mainCamera 主相机。
  * @param wpos 世界空间位置。
  * @param uiNode UI节点。
@@ -24,18 +23,7 @@ export function WorldNode3DToLocalNodeUI (mainCamera: CameraComponent, wpos: Vec
         out = new Vec3();
     }
 
-    mainCamera.worldToScreen(wpos, _temp_vec3_1);
-    _temp_vec3_1.x = _temp_vec3_1.x / cc.view.getScaleX();
-    _temp_vec3_1.y = _temp_vec3_1.y / cc.view.getScaleY();
-    const cmp = uiNode.getComponent('cc.UITransformComponent') as UITransformComponent;
-
-    if (!cmp){
-        return out;
-    }
-
-    cmp.convertToNodeSpaceAR(_temp_vec3_1, out);
-    const targetPos = uiNode.getPosition();
-    out.add(targetPos);
+    mainCamera.convertToUINode(wpos, uiNode, out);
 
     return out;
 }
@@ -45,6 +33,7 @@ export function WorldNode3DToLocalNodeUI (mainCamera: CameraComponent, wpos: Vec
  * Conversion of non-UI nodes to UI Node (World) Space coordinate system.
  * @zh
  * 非 UI 节点转换到 UI 节点(世界) 空间坐标系。
+ * @deprecated 将在 1.2 移除，请使用 CameraComponent 的 `convertToUINode`。
  * @param mainCamera 主相机。
  * @param wpos 世界空间位置。
  * @param out 返回世界坐标。
@@ -60,6 +49,9 @@ export function WorldNode3DToWorldNodeUI (mainCamera: CameraComponent, wpos: Vec
     return out;
 }
 
+/**
+ * @deprecated 将在 1.2 移除，请使用 CameraComponent 的 `convertToUINode`。
+ */
 const convertUtils = {
     WorldNode3DToLocalNodeUI,
     WorldNode3DToWorldNodeUI,
@@ -67,3 +59,13 @@ const convertUtils = {
 
 export { convertUtils };
 cc.pipelineUtils = convertUtils;
+
+replaceProperty(cc, 'cc', [
+    {
+        name: 'pipelineUtils',
+        newName: 'CameraComponent.convertToUINode',
+        customGetter: () => {
+            return CameraComponent;
+        },
+    },
+]);
